Revert "Merge branch 'ajb/rotf' into ajb/rst"
This reverts commit aa65ba90e23f5bb6b87a368b8784c2dccbe698d1, reversing
changes made to 8f805d6f2e1f9f2cd0d3fd65ec209729efc3d8ec.
